lec-notes/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai-whisper
requests
ffmpeg-python
pydantic